Удалена функция короткого вывода

+++ lab2/experience.h
@@ -2,8 +2,8 @@
 
 //АТД стаж
 typedef struct experience {
-	float workingyears; //кол-во реально отработанных лет
+	int workingyears; //кол-во реально отработанных лет
 	int army; //кол-во лет в армии, если сохранялось рабочее место
-	float maternityleave; //кол-во лет в декретном отпуске, если сохранялось рабочее место
+	int maternityleave; //кол-во лет в декретном отпуске, если сохранялось рабочее место
 }Experience;
